+++ lib/src/widgets/nes_icon.dart
@@ -743,6 +743,13 @@ class NesIconCollection {
       '8,8;2,-1;1,0;2,-1;1,0;3,-1;2,0;2,-1;2,0;2,-1;6,0;3,-1;4,0;5,-1;2,0;6,-1;2,0;6,-1;2,0;6,-1;2,0;3,-1',
     ),
   );
+
+  /// Discord
+  late final discord = NesIconData(
+    MiniSprite.fromDataString(
+      '8,8;9,-1;6,0;1,-1;10,0;1,-1;2,0;1,-1;12,0;4,-1;2,0;16,-1',
+    ),
+  );
 }
 
 /// {@template nes_icon}